// Backend/src/model/daftarPermohonan.ts
export interface PermohonanData {
    id_jenis_izin: number;
    id_badan_usaha: number;
    id_status_oss: number;
    id_progress: number;
    id_kabupaten: number;
    id_kecamatan: number;
    id_desa: number;
    tautan_dokumen: string;
    nomor_permohonan: string;
    tanggal_permohonan: string;
    alamat: string;
    modal_usaha: number;
    tanggal_selesai_permohonan: string;
    durasi_permohonan_menit: number;
}